requirements.txt
bcrypt>=4.2.1
certifi>=2024.12.14
cffi>=1.17.1
charset-normalizer>=3.4.0
cryptography>=44.0.0
future>=1.0.0
idna>=3.10
Jinja2>=3.1.4
junos-eznc>=2.7.2
lxml>=5.3.0
markdown-it-py>=3.0.0
MarkupSafe>=3.0.2
mdurl>=0.1.2
napalm>=5.0.0
ncclient>=0.6.15
netaddr>=1.3.0
netmiko>=4.5.0
netutils>=1.10.0
ntc_templates>=7.5.0
paramiko>=3.5.0
pycparser>=2.22
pyeapi>=1.0.4
Pygments>=2.18.0
PyNaCl>=1.5.0
pyparsing>=3.2.0
PyQt6>=6.8.0
PyQt6-Qt6>=6.8.1
PyQt6-WebEngine>=6.8.0
PyQt6-WebEngine-Qt6>=6.8.1
PyQt6_sip>=13.9.1
pyserial>=3.5
PyYAML>=6.0.2
requests>=2.32.3
rich>=13.9.4
ruamel.yaml>=0.18.6
ruamel.yaml.clib>=0.2.12
scp>=0.15.0
six>=1.17.0
textfsm>=1.1.3
transitions>=0.9.2
ttp>=0.9.5
ttp-templates>=0.3.7
typing_extensions>=4.12.2
urllib3>=2.2.3
yamlordereddictloader>=0.4.2
